Title: 가장 많이 겹치는 구간 2, Time: 17ms, Memory: 0MB, Status: Runtime Error - Codetree

+++ 260314/가장 많이 겹치는 구간 2/section-with-maximum-overlap-2.cpp
@@ -6,9 +6,9 @@ int n;
 long long x1[100004], x2[100004];
 long long arr[100004];
 //1000 000 000
-int mx = INT_MIN;
+long long mx = INT_MIN;
 long long ans = INT_MIN;
-long sum;
+long long sum;
 int main() {
     cin >> n;
     for (int i = 0; i < n; i++) {
